feat(app): implement Fastify server with Swagger documentation and health checks
